WIP
This teaches FieldLoader to compile a delegate at runtime for loading the target class.

This delegate is about twice as fast as using reflection, and allocates less as it can avoid boxing.
The downside is that compiling this delegate is about 1000x the cost of executing it. This adds a second or two of extra time on first load.

After first load, we'll benefit from the marginally faster delegate when loading these classes again, but the up front cost is pretty high.

The best solution here would be a source generator to push this to compile time instead.